remove social account service (#53)
